requirements.txt
scheduler
pyTelegramBotAPI
pysocks
numpy